requirements.txt
pandas
sqlalchemy
pyodbc
psycopg2-binary
azure-storage-blob
python-dotenv
nbformat
jupyterlab
